requirements.txt
decorator==4.0.11
Django==1.10.5
django-celery-beat==1.0.1
django-cors-headers==2.0.2
django-environ==0.4.1
django-extensions==1.7.6
django-grappelli==2.9.1
djangorestframework==3.5.4
entrypoints==0.2.2
enum34==1.1.6
et-xmlfile==1.0.1
functools32==3.2.3.post2
future==0.16.0
gunicorn==19.7.1
html5lib==0.999999999
ipykernel==4.5.2
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jdcal==1.3
jieba==0.38
Jinja2==2.9.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.0
jupyter-console==5.1.0
jupyter-core==4.3.0
kombu==4.0.2
line-bot-sdk==1.1.0
pyyaml==3.12
